When opening a new stack, make sure the scale is changed about the right point
The recent changes to zoom behaviour broke opening a project
about an exact point (e.g. via "URL to this view"), since
zooming now takes place preserving the position of the point
the mouse pointer was last at - when opening a stack for the
first time this value will not be anything sensible.

Thanks to @perlman for reporting this bug.

Closes #50