An attempt to fix issue 53
There may be other instances of this problem, where a variable
is not declared with "var" and its name is also the id of
an element in the page:

   http://msdn.microsoft.com/en-us/library/ms536437%28v=vs.85%29.aspx#4

... but for the moment, I'll just change this instance, to see
whether this fixes it for @jovo.  If that works, I'll track down
the others.  (All variable should be declared with var, of course,
and this will be picked up by JSLint - we haven't got to init.js
yet, though.)