Change how treenodes and lines between them are drawn
This is an attempt to simplify the logic of picking colours
for treenodes and the lines connecting them.

There's a possible performance concern here, in that
the recolorAllNodes sets the color for all visible nodes.
This doesn't currently seem to be a problem, but might
be in the future.

This closes #63