Add basic support for automatic database migrations
When the DB object is first created, check that the schema version
in the 'settings' table is the same as the most recent migration
that has been defined in migrations.php.  If it isn't (either the
schema version couldn't be found, or the database is out of date)
then go through each migration that hasn't yet been applied and
try to apply it.  Finally, update the schema version in the settings
table with the value of the last migration.

We want this to work with existing databases, so if the schema
version couldn't be found initially, try to apply every migration
but ignoring errors.  (There will probably be lots of errors,
since many of the tables will already have been created.)

The only migration defined is one that assumes that your database
is in the state represented by docs/master.sql from 6574a2e66