Fix for coordinates sometimes being wrong in Internet Explorer 8
The getMouse function in ui.js uses offset[XY] on IE 8 and
layer[XY] - 1 on Firefox.  Unfortunately, offset[XY] is relative
to the circle SVG element if you mouse over a node, which creates
positioning bugs in IE.  This commit should fix that problem
in mousemove, although there may be other instances of the same
problem.