Change the behaviour of ui.getMouse()
The calculation of offset[XY] returned from getMouse
didn't work cross-browser, since the event's offset[XY] and
layer[XY] are defined very differently - this was exposed when
you would mouseover an SVG element in Internet Explorer.

Another change here is to report in the returned [xy] the
coordinates relative to the top left of the document, not the
top left of the screen.  This is more useful, and doesn't affect
the only use of these properties in the code.