Switch to a custom postgresql_psycogpg2 adaptor
One of the ways in which the postgresql_psycopg2 adaptor doesn't
work with inherited tables is that the SQL it uses to get the
ID of the most recently inserted row in a table doesn't work:
pg_get_serial_sequence requires the sequence to be owned by that
table, and with inherited tables there is one sequence being used
by the primary keys of multiple columns.

Instead, I've overridden last_insert_id to find the default value
of the primary key for the table and parse out the sequence name
from that.  This isn't particularly pretty, but it seems to work
and be the least intrusive change.