Fix for coordinates sometimes being wrong in Internet Explorer 8
The getMouse function in ui.js used offset[XY] on IE 8 and
layer[XY] - 1 on Firefox.  Unfortunately, offset[XY] is relative
to the circle SVG element if you mouse over a node, which creates
positioning bugs in IE.  This commit should fix that problem
by allowing you to explicitly specifying the element that
offset[XY] should be calculated with respect to.

Another change here is to report in the returned [xy] the
coordinates relative to the top left of the document, not the
top left of the screen.  This is more useful, and doesn't affect
the only use of these properties in the code.