Clarify getActiveNode in SVGOverlay, and make sure it only null or a node
It may not be clear that if the active node is out of view,
this function will not return the node.  Clarify that in the
comment.

Another problem fixed here is that if that active node is out of
view, undefined is returned - it's probably clearer if it
consistently returns null for failure.