Make the 'P' (go to parent) keyboard shortcut work (most of the time)
There are situations where 'P' will claim that the active node is
already the root node, when in fact it is just that the parent node
was not in the set of nodes fetched.  With the z range that we
currently use for fetching, this means that if the parent is more
than one section away (or out of the x / y range) you will get this
error.  Probably the solution to this is to make sure that we
fetch all the nodes of the active skeleton no matter where they
are in the stack.