Improve node deletion behaviour
It used to be the case that deleting a node when a distant node
was active would cause the parent of the deleted node to become
the active node, which was surprising and annoying.  Now the
parent of the deleted node only becomes active if the node was
active when deleted.