Add getMap to the DB type, to fetch all classes or relations in one query
Various PHP files being with multiple queries to return the class
IDs corresponding to class names for a particular project.  This
is obviously sub-optimal, since we might as well fetch all of these
in one query and return an array that maps class names to class
IDs.  The same method works for relations instead of classes if you
pass type 'relation' instead of 'class'.