There should be no error when adding a distant post-synaptic terminal
If you created a connector, made the connector active, moved two
slices away, and then used shift-click to create a post-synaptic
terminal, there would be an error.  This commit avoids that error,
but isn't a complete solution - the error arose because of trying
to update and redraw the connector, so that an arrow is correctly
drawn to the new terminal.  There needs to be further work to
draw arrowheads pointing to post-synaptic terminals that are
connected to out-of-view connectors in the general case.