Order elements in the object tree by creation_time
Previously these were ordered by edition_time in descending order,
which meant that the most recently edited elements went to the
top of the list.  In the absence of any functionality allowing us
to reorder the list manually, it's better to order by
creation_time, which will match the order that was found in
TrakEM2.