Connector table: also show connector with no partner treenodes
This could be done in a single query if we used LEFT OUTER JOIN,
but the query is probably complex enough to understand already.
Instead we do an extra query which just tries to find all
connectors, and add any which weren't found previously to the
results.  Missing data is replace with the empty string.

This also means that the behaviour when you double-click in the
connector table is slightly different - if there's no partner
treenode, then the connector itself is activated.