network-api.js: fixed error in the listing of connectors in a skeleton.
Before, a skeleton node could be listed as having a single connector.
Now each skeleton node may be listed as having an array of one or more connectors.

For convenience, Skeleton got new functions preConnectors and postConnectors,
which return flat arrays of unique connectors for the skeleton.


Thanks to Mark Longair for identifying the error.