Add a script to rename directories if a missing layer was really a removed duplicate
This is quite specific to a situation we encountered, where
a layer that I regarded as missing was really a duplicated
layer that was removed without adjusting the z coordinates.
This script contains in comments at the top the corresponding
SQL that should be used if you have to rename the layers in
this way.