A quick fix for the "URL to this view" links not working
The commit 675d047ff8bd6fb79a0ad3c29f25d0d15c20b090 broke
the following of URLs to a particular view.  This was because
the use of PHP's usort to sort the array of projects loses
the old array keys (which where the project IDs) and
renumbered each project array to 0, 1, 2, etc.  The logic
in handle_updateProjects still assumed that they keys are
project IDs and then failed to set keep_project_alive to
true.

The simplest fix (introduced in this commit) is just to add
a pid member to each project array, and change
handle_updateProjects to use that field for comparison.

Fixes #208.