Make the return values from handleKeyPress() consistent
The actions used to return false when they had successfully
dealt with an event, while the code that called it required
them to return true in that case.

This also incidentally fixes #178.  Now we only return true
on handling the keypress with the tracing tool if control
was *not* held down.