Using 'Alt' as a modifier is much less problematic
If you try to detect the 'Shift' modifier for the number
keys 1 to 5, all sorts of problems arise.  For example
Shift-2 is sometimes interpreted as 222 (on UK keyboards)
and under Firefox (on a UK keyboard) Shift-3 (£) is sent
as 0.  It's easier to use 'Alt' as a modifier, since
Alt-2 wouldn't generate those characters.