Cropping tool: add new Django back-end
The actual cropping is now done as part of a python script and can be
called through a Django view. The image manipulation is done with the
help of GraphicsMagick (pgmagick), because the simpler PIL doesn't
allow the creation of multi-page images. This feature is needed for
exporting image stacks and GraphicsMagick supports that.

To process a cropping job one must be logged in. If that is the case
the URL (relative to django) looks like this:

/<pid>/stack/<sid>/crop/<x_min>,<x_max>/<y_min>,<y_max>/<z_min>,<z_max>/<zoom-level>/

Like starting the cropping job, the download is also done with the help
of a Django view:

/crop/download/<file_name>/

It is required that the Django settings (settings.py) have a variable
called TMP_DIR defined. Cropped images will be stored there. Like with
the older cropping tool, cropped images are allowed to stay there for
two weeks. If the temporary folder is accessible and writable is checked
for before executing the job. A small sanity chech of the parameters is
also done.

One is allowed to include non-data (i.e. backgound) regions in a crop,
but only if parts of the stack could be seen. Creating pure background
images will be refused.