Cropping tool: do the actual cropping asynchronously
As cropping requests could take some time due to various reasons (e.g. size
of sub-stack, server load) they should be processed in the background. This
is implemented using Celery, a asynchronous tash/job queue. This requires a
celery daemon (celeryd) to be running. This can be started manually with

    manage.py celeryd

or alternatively be integrated in the system sertup to start automatically.