Cropping tool: export some meta data into the created sub-stack image
Information about the resolution and the units used is important to
set the cropped sub-stack into picture. The EXIF tags XResolution,
YResolution und ResolutionUnit (None) are exported. Also, ImageJ specific
meta data is added to allow conveniert access to unit and display
information from within ImageJ.

Unfortunately, no Python module was working for this (mainly due to a
bug in exiv2 concerning multi-page images) and a seperate tool had to
be used: exiftool. Like added to the INSTALL document, this has to be
installed.