Add menu item "Projects > Current" to list stacks of current project
If a project has more than one stack a new menu item is created when a stack
of the project is opened. This item ("Project > Current") is a shortcut for
finding the current project in the "Project > Open" menu. If the number of
projects increases, this can be very handy.