Fix the selection of "Isolated synaptic terminals"
The "Isolated synaptic terminals" group in the object
tree is treated specially.  Normally, only skeletons
that contain more than one treenode are returned as part
of this group, to avoid opening a group with thousands
of children.  However, if we select a treenode from
a single skeleton whose neuron is in that group in the
overlay, then we want to display it in that group
nonetheless, opening the folder and selecting as with
any other treenode.

A previous commit (c01445b6) introduced the special
behaviour for "Isolated synaptic terminals" and also
ensured that if a request for elements of the object
tree was made because of a treenode that descends
from "Isolated synaptic terminals", that treenode
and its ancestors would be included in the results
in any case.

This commit changes the behaviour of openTreePath
so that if it fails to find a node that it wants to
open, it will try refreshing (once) before trying
again, so that the node might be refreshed.