Introduce the concept of a tile_data_source.
In order to use arbitrary TileSource server, such as for
dynamically generated tiles with POST request, a simple
type scheme has been introduced to distinguish between
two types. Further description in data-model.lyx.

The reason for this change was to become more flexible
in the tile naming URL, and e.g. retrieve tiles from
POST requests with a set of parameters specifying the
tile location.

This is for a proof-of-concept to show the integration
with the Ilastik Volumina data slicer.