Add labelupload URL to stack
The labelupload URL can be set by TileServer source type 2 and
is used to upload labeling pixel for segmentation using the canvas
tool. The tile source type 1 (old CATMAID tile scheme) does not
support this functionality in and this URL is not reflected in the
CATMAID database.

The labelupload_url is used in the Canvas Tool when clicking
'Send labels' using the proper field of view into the stack for
indexing into the stack data array.