Only store the parent ID in atn, not a reference to the object
The problem with having a reference to the parent node in
atn (the active node) is that on scrolling the canvas, that
node may be reused for a different node.  Instead, just store
the ID, and when you need the parent of the active node,
find it in the nodes array.

This affects the go-to-parent keyboard shortcut, so this
commit also refactors that code, and removes a redundant
copy.