Avoid a possible loop which would hammer node.list.php
The redraw method relies on stack.old_z being updated to
tell whether an updateNodes needs to be performed.  However,
this is onloy updated in the stack's redraw method, and
handle_updateNodes was calling the overlay's redraw directly.
This could cause a loop where the overlay kept redrawing
and calling updateNodes.

Instead, call stack.redraw() from the handle_updateNodes
callback.