Refactor node and line drawing so it is easier to customize
Previously, it was difficult to work out how the drawing of
nodes and lines was controlled.  This commit hopefully
simplifies this by introducing the following methods
in overlay_node.js:

   displayConnector
   displayTreenode
   displayBetweenNodes

... which should return true if the structure should be
drawn.  The implementation of these in this commit is
to draw connectors and treenodes if they are within
half a slice of the current slice, while lines or arrows
between nodes will be drawn only if one of the two
would be drawn.

We will probably want to customize this behaviour, but
at least this commit makes it simple to do so.