Also fetch all connectors up to 4 slices on either side
This is so that if a treenode which represents a pre- or
post-synaptic terminal is several slices away from the
connector, an incoming arrow will still be drawn.

Since there are many fewer connectors than treenodes
(10^4 vs 10^6 on zanzibar) this shouldn't be too much
of a performance hit during tracing.

The other alternative solution would be to join with the
treenode_connector table when fetching treenodes in the
first place, but the performance impact of that would
be much greater, we believe.