Added Review System for Skeletons
After fetching segments of a skeleton in the review widget, you
can start reviewing by clicking on the Review button in each row.
You jump to the start of the segment. Then you can move back and
forth the segmenet with the 'Q' and 'W' keys. After hittig the end
point of the segment, the review widget is updated. You can now
start reviewing another segment. If you are finished with reviewing
all segments of a skeleton, click 'End Review' button to release
the review state.

During the reviewing, you are free to add nodes (which have to
be reviewed by somebody else), or add tags to nodes etc.