Added wiki_base_url to project table
This allows integration of e.g. a (semantic) mediawiki
installation in the Neuron Catalog. I.e. for a given
neuron in the Neuron Catalog View, an URL is generated
using this base URL and the neuron name to display
the corresponding wiki page in an IFrame.

This requires setting $wgEditPageFrameOptions = false;
in the semantic media wiki installation to allow inframe
editing.