Replace search.php
Note that the replacement does not exhibit the exact same behavior as
the PHP function. We return integers instead of strings of the integers
(for coordinates and ids). This should not be a problem, as the only
place that uses search.php wraps all these in parseInt(), which returns
the same thing when passed integers or strings of integers.

Diff for urls.py might look odd, but it is just re-ordering some lines
and adding a line for search.