Add script to grant Django auth group access to CATMAID projects
This script will ask for a Django authentication group name and
allows to select permission options for CATMAID projects to grant
those permmisons to selected CATMAID projects. The selection of
projects is determined by asking whether to only look at public
projects and by allowing to run a regular expression filter on
the image bases of the stacks associated with a single project.
If all stacks match the condition, the project is selected. By
default, all stacks match.

At least when a lot of projects are in use, this script allows
for an easy permission management based on Django's auth groups.