Make building of projects data structure faster
Especially when there are many projects, this will considerably
speedup the the creation of the projects data structure. When
tested with about 900 projects, the time needed went down from
about 45 seconds to 0.4 secconds. This structure is created when
CATMAID is loaded. Therefore, the site loading time decreases by
the same amount.