Separate out function to add edit and catalogue info to project lists
With the help of this method any list of projects can be replaced
by a list of wrapped projects where each wrapped project keeps
information about whether the project is editable and catalogueable.

This additional information is e.g. needed to create the JSON object
returned by /projects. Since the same data is also needed for data
view rendering, it is factored out to be reusable.