Make project table managed by Django. Update documentation.
When installing CATMAID from scratch with an empty database,
the migrations and the manage.py syncdb command were circularly
dependent on each other, preventing to setup the database tables.
Now, the project table is managed by Django and created with syncdb
and the migration does not create the project table and associated
keys.

This has to be done similarily for the stack and project_stack
tables in the future.

An additional path for the Apache specifiying the CSS for the
admin interface is added to the Django configuration script.