Data views: Add display of stack comments to templates
The avaiable data view types display stack comments in different ways:
In a project list a comment is added as text to the right of the stack
link. The images in a project table data view are now placed in a
figure-tag and get the comment as caption and as image tool-tip. When
displaying text, the comment is visible in the tool-tip, too. Within the
project tags table, images and text links get a comment only as tool-tip.