Importer: add optional URL, zoom level and file extension in YAML
If a stack in a project file now contains a 'url' field, it is used as
the stack's image base. If done so, the fields 'zoomlevels' and
'fileextension' need to be available as well. If no 'url' field is
present, the importer looks for the 'folder' setting. If there are then
still 'zoomlevels' and 'fileextension' fields around, those will be used
as well. The importer will retrieve this information only if one of them
is missing.