Add X-Requested-With header for request send with request queue
According to [1] the 'X-Requested-With' header is used to identify AJAX
requests. If it is an AJAX call, the header field's content should be
'XMLHttpRequest'. When the request queue is used to create a request,
an AJAX request is issued. Therefore it should be marked as such.

Django's request.is_ajax() method checks this header field and returns
true if it is present and its value is 'XMLHttpRequest' [2].

[1] http://en.wikipedia.org/wiki/X-Requested-With#Requested-With
[2] https://docs.djangoproject.com/en/1.4/ref/request-response/#django.http.HttpRequest.is_ajax