Skeleton treenode review system: create chunks continuous across branch nodes, and always traversing from end node towards root.
The longest path will traverse from the furthest end node to the root node.
The second longest will traverse from an end node to a branch point that
has already been seen. And so on.

The desired effect: create less sequences to inspect which minimizes context
switching (changing of location in the image series), and also not creating
a long list of short sequences between branch points.