Remove try/except block from cropping imports
The try/except block was meant to make things easier for people who
don't want to or don't need to use Celery. The idea was to treat
import errors of it silenly. However, since there were multiple
imports in this try/except block, a failed import of the first
import would cause the second (and third) not to be even tried. To
not further complicate things, I just removed the try/except block
and expect Celery to be configured. This shouldn't be a problem
since the default settings file already provide a minimal working
configuration with djcelery.