Use default template context processors of Django
Template context processors provide templates with a basic context like
the path of static files. Up to now the default processors were not (!)
used. That confused some Django applications like django-guardian which
couldn't refer to some static files because of this. This commits
ensures that Django's default context processors are in use and allows
to easily add custom ones.

Additionally, an old and not used custom template context processor was
removed. Now there are only the defaults processors in use.