Add empty migration that has changes of both 0002 migrations
Since one migration got merged in, both 0002 don't know about each
other. This can easily be handled with

  ./manage.py migrate catmaid --merge

However, to let new migrations refer to the correct database schema,
this new empty migration incorproates the changes of both 0002
migrations and contains an consistend database schema. It was created
with

  ./manage.py schemamigration --empty catmaid merge_models

To read more about this, have a look at the Team Workflow section of the
South documentation:

  http://south.readthedocs.org/en/latest/tutorial/part5.html