Ensure that a UserProfile is created for the anonymous guardian user
There is a problem in that guardian creates an anonymous user when
its initial migrations are run, but a corresponding UserProfile will
not be created.  This commit introduces code to monkey-patch
guardian.management.create_anonymous_user to make sure that the
UserProfile is created.

This solution was worked out with Tom Kazimiers.