Ontology editor: display the known root names
CATMAID's ontologies are graphs where the nodes are classes and edges
are relations. The ontology editor displays this graph within a tree
structure and needs therefore a starting point. This starting point is
defined by a root class. Currently, only the root class named "root" is
known to the ontology tool. However, this can easiliy be extended. This
commit adds a display of the known root classes to the ontology editor.