Ontology editor: don't allow creation of new Cs/Rs in ontology tree
It can be very confusing if the context menu of the ontology also allows
the creation of new classes and relations. Therefore, these options are
removed with this commit. Classes and relations have to be created in
the class and relation trees before they can be used in the ontology
tree.