Classification editor: check permissions in templates
All classification templates now render only editing/changing options
if the current user has the 'can_annotate' permission for the current
project and the current workspace.

In the front-end the attachment of one call-back method had to be
adjusted. It will now only be used, if the target container is actually
present (which it might be not due to limited permissions).