Classification editor: remove jsTree UI plugin to make ROI links work
The UI plugin of jsTree adds hovering and selction functionality to
nodes. However, it doesn't allow any click event to pass through to the
actual node. This in turn is needed to make links in a node, e.g. to a
ROI, work. Therefore, the UI plugin has been removed for now.