Return the actually used stack after added a stack to a project
To display stacks of a project in the front-end, stacks are added to the
project object. In case one tries to add a stack that is loaded already,
the project keeps using the already loaded stack. If the stack is not
present, it will be added. This commit adds a change to make the
actually used stack available to the caller, by returning a reference to
it.