Make sure the correct stack is used after a stack is added
In case a stack with the very same ID is already present in a project,
this existing stack will be used and no new stack will be added. Since
the project opening code continues using the stack after adding it, this
commit makes sure that the correct stack instance is used after it is
tried to add a stack.