Call a tool's unregister() method when focus on a stack is lost
This change re-enables the notification of a tool of being unregistered
for the current stack. It brings it back to how the call structure was
intended originally. This behaviour was deactivated to reduce the number
of redraw operations, but no new redraw() and resize() calls are
made through this commit, as long as a tool's unregister() method
doesn't issue them. This commit fixes #509.