Put inital data view types data into fixture
The data in this fixture will be loaded on every manage.py migrate
call (because of its name). Changes to the data view types available
can now be made to the data in the fixture file instead of creating
a new migration. This makes the handling of default data cleaner.