Don't attempt to create super-user during syncdb
All users in CATMAID have a user profile attached and every new user
gets automatically one created. This, however, does only work if the
model tables have been created already. This is not the case when syncdb
asks if it should create a superuser. This commit disables this question
and removes the note in the manual about saying "no" to it. Still, a
superuser has to be created manually.