Clustering: allow sub-path as features
When selecting the "Use sub-paths as features" option, all sub-paths
starting at the root on a way to a leaf will be added as features, too.
This means a path doesn't need to end with an ontology defined leaf, but
it can also end in the middle to be respected by the clustering.