Allow HTML code in notes/comments of menu items
If a menu item has a note/comment associated, it now is renderd as-is.
This means they now can contain HTML code (e.g. links) and let the
browser render it. Each note is embedded in a div of the class
"menu_item_note" to make the display easier adjustable. This commit
fixes #536.