Retrieve stack menu info when needed instead of caching
In three spots, the stack menu information for the current project was
needed. Up to now cached information was used to get this data. This
commit changes those spots to request the stack info right when needed.
This simplifies the code a bit as the cache doesn't need to be kept up
to date anymore.