Data views: only close current project when loaded from menu
Up to now the loading of a data view (the front page) caused any open
project to close. This commit changes this behavior to only close the
current project (and its windows) when the user opens a data view
through the data view menu (appearing when hovering the "Home" link).

Projects loaded with the help of a "URL to this view" link, now won't
get closed immediately anymore because of loading of the default data
view. This is related to isuse #537.