Make sure initial stacks get first loaded after login completion
The URL to a CATMAID instance can contain information about which stacks
to load at startup. Such a URL can be created with the help of tho "URL
to this view" link. The loading of these stacks, however, requires the
login procedure to have finished. This commit makes sure this is the
case by loading initial stacks in a callback method passed to the login
command. Without this, the handle_updateProjects() method would destroy
the newly created project right away. The callback is instead executed
*after* the potential deletion of an existing project. This is related
to issue #537.