Issue project update after logout
Like after a login, updating the available projects makes sense after a
logout: The permission change most likely causes a change in the list of
available projects.

This also fixes the alleged hang afer a logout, because updateProjects()
calls ui.releaseEvents(). This wasn't called before, despite the fact
that ui.catchEvents("wait") is called in logout().