Add admin tool "Tag Based Classification Graph Linker"
The tool introduced by this commit adds a new custom admin view with
which one can link classification graphs to projects in an automated
fasion. This is done based on the attached tags of a project. The tool
finds tag sets and offers to unify the linked classification graphs
or projects within those tag sets.