Add optional importer step to link suggested classifications
This commit introduces a new optional option called "Check
classification links". When checked a new importer step is added after
the project selection. This step suggests existing classification graphs
to be linked with the new project(s). The suggested graphs are project
independent and have the the same tag set (or a super set of it) as the
new project(s). The user is then free to mark those classification
graphs for automatic linking after the final confirmation step.