Clustering: add and use PL/pgSQL function to find features in ontology
A feature path is a set of nodes that start from the root of an
ontology. The DBMS-function added by this commit, provides a very fast
way to get all feature paths of an ontology. Just as the Python
version, it allows to specify if only root-leaf paths should be added or
if all intermediate endings should be added as well.

This PL/pgSQL implementation needs only about 5% of the time of the
Python implemenation (which uses Django's ORM and produces many
queries).