Show more reasonable ID in a CICI link's text representation
With respect to a classification root node, class_instance_b of a
ClassInstanceClassInstance (CICI) link is the actual root class instance
of a classification graph. Its class_instance_a would refer to the
linked project. Since there can be many CICI links to the actual root,
but from different projects, it is makes sense to not show the link's
ID, but the root's.