Fix problems with overlays in stack window
If the slider for an overlay is set to 0 (i.e. smaller than 0.02 to
be precise) - the tilelayer is removed from the DOM and kept only
as a reference. Tiles are then not fetched. In the Django admin
interface, set the default opacity to 1 (or a value in the range 0
to 100) to isolate the tile layer by default when loading the stack.
This allows now to have many overlays attached to a stack, but only
fetch the tiles for the overlay you turned up the opacity value.