webglapp: reduced memory footprint dramatically
* By reusing Vector3 instances when creating edges of the skeleton
(used to have 2 new Vector3 per edge in the skeleton, and now there is 1).
* By reusing Vector3 for the skeleton node part of the connector edge.
* By not storing nodeProps and connectorProps, which were an array of arrays,
each with the skeleton node position and related IDs. Now the ids (node_id,
user_id, and reviewer_id) are stored as fields in the Vector3 of the geometry.
No new data is loaded; instead, the existing information in the geometry
is mined to create the graph or recolor the nodes or recompute the restricted
connectors.