Multiple selection widgets and multiple instances of widgets.
The Connectivity, 3D View and Graph to act as a source of selected
skeletons like the Selection Table.

Each selection widget has its own concept of what is selected:
 * Graph: selected nodes
 * 3D View: visible skeletons
 * Connectivity: seed skeletons plus any selected in the pre/post tables.

Each of the selection widgets extends InstanceRegistry and SkeletonSource,
and can load skeletons from each other or the active skeleton in the overlay.

Each selection widget exports colors as it pleases. The connectivity widget
makes up the colors depending upon the condition of being a seed neuron
or pre or post or both. The active skeleton is always yellow.

The Skeleton Analytics merely extends InstanceRegistry to enable multiple
instances.