Neuron annotations: make neurons A and annotations B for CICI links
A ClassInstanceClassInstance links class instance a with class instance
b through a relation. Up to this commit annotations were treated as
class instance a and neurons as class instance b. This commits reverses
this, because it is more natural to read "a neuron is annotated with an
annotation" (given <class instance a> <relation> <class instance b>).