Neuron navigator: add node types and node selection logic
Nodes represent pages in the navigator. They provide content and know
there parents. Internally they are represented as sub-classes of the
NeuronNavigatorNode class. Currently, there is a home node, a users node
and an annotations node available. However, these are merely the frames
of what they should become.

Since every node knows its parent, it is easy to create the navigation
bar which represents every node in the path from the current one to the
home node: A node constructs it by asking its parent for the the bar and
adds itself at the end.