Neuron navigator: allow requests to user-list API to ignore users
When using the POST HTTP method to request a user-list, it is now
possible to add a list of user IDs to ignore in the response. This is
needed for the user filters within the neuron navigator. There is no
change for requests made from existing code, i.e. regular GET requests
are still supported.