Allow parameterization of new window creation
This commit adds a new parameter to the window manager's create() method. It is
passed on to the method actually creating the requested window. This parameter
allows to get data into the otherwise isolated window creation process. However,
a window creating method doesn't need to make use of this.

This functionality may be used if a widget should be cloned.