Neuron navigator: add button to duplicate widget
The button is placed in the upper right corner of the navigator widget. When
clicked, it will instruct the window manager to create a new window, based on
a new navigator instance it created. The current node will be deep-copied and
will therefore clone all it's parents. All new nodes are then linked to the new
navigator and the current one is selected.