Neuron navigator: add one-level annotation filter
The added filter works for neurons, users and annotations. It's influence is
limited by one level. If neurons are selected after this filter, only neurons are
displayed that are annotated with the selected annotations. A user list selected
right after this filter will show only show users that have used the selected
filter. If annotations are shown after this filter, only annotations are shown
that are annotated with the selected annotations (i.e. the selected annotation
is a meta annotation).