Neuron navigator: add skeleton statistics to neuron node
At the moment, the neuron node shows the following statistics for each skeleton
in a data table: number of nodes, number of branches, number of end nodes and
number of open end nodes. End nodes are nodes that are tagged with either
"ends", "uncertain end", "not a branch" or "soma".

To create the statistics the compact-json representation of a skeleton is
fetched. Based on this an Arbor instance is created to have access to convenient
traversal and filter algorithms.