Neuron navigator: support co-annotations
Filters like user or annotation provided by the navigator are only valid for one
level. Co-annotations differ in that respect. If an annotation has been selected
and one chooses a co-annotation right after that, it adds to a combined
annotation filter. That is, user or neuron selections after will honor not only
the annotation chosen before, but also the co-annotations that have been
selected after it.