Skeleton splitting: add annotation based splitting
When splitting the front-end and the back-end now require that one part
maintains all previous annotations. The first part is referred to as 'over' and
the second one is called 'under'. The under skeleton will now be a model of a
newly created skeleton which also gets all the under annotations.