Neuron annotations: add new permission testing function
The method can_edit_class_instance_or_fail provides a similar test to the
existing can_edit_or_fail method. It is, however, based on th new annotation
system. A class_instance (e.g. a neuron) can be locked, i.e. it is annotated
with the annotation named 'locked'. A user is only granted permission to edit a
class_instance if it is a) not locked or b) if it was locked by the user or by
a user of who's group (s)he is part of.