Circuit Graph Plot: created new dimension "Graph partition (cell types)" which combines:
* for disjoint networks, the second and third eigenvectors
by getting the non-zero values.

* for single-component networks, combines the third and fourth eigenvectors
according to the sign in the second eigenvector. The positive values
in the second eigenvector are then picked from the third or the fourth,
depending upon which has the highest standard deviation for these values.

The effect is that neurons that are connected to similar sets of neurons
in similar ways will end up together in this dimension, which effectively
identifies cell types by connectivity patterns.